+++ CHANGELOG.md
@@ -10,8 +10,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - `AgentInfo` to represent the agent's information during the handshake process.
-- `server::handshake` implements the application-level handshake process for the
-  server after a QUIC connection is established.
+- `client::handshake` and `server::handshake` implement the application-level
+  handshake process between the client and the server after a QUIC connection is
+  established.
 
 ## [0.1.2] - 2024-03-25